# vertexfilter: Optimize vertex filters using SSE2