This results in dramatic performance improvements for native builds:

master:
filter: oct8 74.12 ms (0.40 GB/sec), oct12 74.17 ms (0.80 GB/sec), quat12 75.14 ms (0.79 GB/sec), exp 4.56 ms (13.07 GB/sec)

filter-sse:
filter: oct8 13.32 ms (2.24 GB/sec), oct12 13.80 ms (4.32 GB/sec), quat12 19.66 ms (3.03 GB/sec), exp 4.57 ms (13.06 GB/sec)

The change also makes quat scalar filter slightly faster and simpler by not using a lookup table.